requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
openai>=1.40.0
pydantic
numpy
